Circular Dependancy [ FIXED ]

+++ Chili Framework 2016_PooGame/Engine/Poo.h
@@ -5,16 +5,16 @@
 class Poo
 {
 public:
-	Poo(int Lx, int Ly, int LvelX, int LVely);
+	Poo(float Lx, float Ly, float LvelX, float LVely);
 	void Update();
 	void ProcessConsumption(const Dude& dude);
 	void Draw(Graphics& gfx) const;
 	bool IsEaten();
 private:
-	int x;
-	int y;
-	int velx;
-	int velY;
+	float x;
+	float y;
+	float velx;
+	float velY;
 	static constexpr int width = 24;
 	static constexpr int hieght = 24;
 	bool pooiseaten = false;
